+++ applications/fishing-map/src/types/vendor.d.ts
@@ -1,1 +1,3 @@
 declare module 'intersection-observer'
+declare module 'auth-middleware'
+declare module 'auth-middleware/*'