+++ packages/layer-composer/src/layer-composer.ts
@@ -1,4 +1,4 @@
-import { Style, AnySourceImpl, Layer, SymbolPaint, AnyPaint } from 'mapbox-gl'
+import { AnySourceImpl, Layer, AnyPaint } from '@globalfishingwatch/mapbox-gl'
 import Generators from './generators'
 import { flatObjectArrays, layersDictToArray } from './utils'
 import {
@@ -75,6 +75,7 @@ class LayerComposer {
       layers.filter((layer) => layer.metadata).map((layer) => [layer.id, layer.metadata])
     )
     const metadata = {
+      generatedAt: Date.now(),
       layers: metadataLayers,
       temporalgrid: layers.find((layer) => layer?.metadata?.temporalgrid)?.metadata,
     }