+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-types
 
+## 2.6.0
+
+### Minor Changes
+
+- af016529: aoi id and dataset schema types
+
+## 2.5.1
+
+### Patch Changes
+
+- include public in workspace types definition
+
 ## 2.5.0
 
 ### Minor Changes